refactor pt 3: split io.py god-object into a package
- split utils/io.py (800 LOC) into utils/io/{writer,restart,debug,cli}.py
- writer: hdf5 + initialise; restart: read_input/sim_restart/fn_gen; debug: the debug writers; cli: get_args/init_logger/mkdir_p
- re-export public names from io/__init__ so every 'from ..utils import io' site is unchanged
- drop dead vortz/vorty/dpress_dim (dpress_dim also used removed np.complex)
- proven inert: io writes the golden-master h5; FAST gate bit-identical at tol 0
